% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imf_get_datastructure.R
\name{imf_get_datastructure}
\alias{imf_get_datastructure}
\title{Retrieve the datastructure definition for an IMF dataflow.}
\usage{
imf_get_datastructure(
  dataflow_id,
  progress = FALSE,
  max_tries = 10L,
  cache = TRUE
)
}
\arguments{
\item{dataflow_id}{The ID of the dataflow to retrieve the datastructure for.}

\item{progress}{Logical; whether to show progress.}

\item{max_tries}{Integer; maximum retry attempts.}

\item{cache}{Logical; whether to cache the request.}
}
\value{
tibble::tibble(
dimension_id = character(),
order = integer(),
concept_id = character(),
concept_name = character(),
role = character(),
codelist_id = character(),
codelist_agency = character(),
representation_type = character(),
stringsAsFactors = FALSE
)
}
\description{
Retrieve the datastructure definition for an IMF dataflow.
}
\examples{
\dontrun{
imf_get_datastructure("GFS")
}
}
